moved monsters to the correct act
