+++ src/main/java/chronometry/monsters/exordium/FungiBeastPatch.java
@@ -1,5 +1,6 @@
-package chronometry.monsters;
+package chronometry.monsters.exordium;
 
+import basemod.ReflectionHacks;
 import chronometry.patches.AbstractMonsterPatch;
 import chronometry.IntentData;
 import chronometry.effects.AttackEffect;
@@ -7,38 +8,38 @@ import chronometry.effects.BuffStrengthEffect;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.monsters.AbstractMonster.Intent;
-import com.megacrit.cardcrawl.monsters.exordium.LouseNormal;
+import com.megacrit.cardcrawl.monsters.exordium.FungiBeast;
 
 import java.util.ArrayList;
 
-public class LouseNormalPatch {
-    @SpirePatch(clz=LouseNormal.class,
+public class FungiBeastPatch {
+    @SpirePatch(clz=FungiBeast.class,
                 method=SpirePatch.CONSTRUCTOR,
                 paramtypez = {float.class, float.class})
     public static class InitMoves {
-        public static void Postfix(LouseNormal __instance, float x, float y) {
+        public static void Postfix(FungiBeast __instance, float x, float y) {
             ArrayList<IntentData> moves = new ArrayList<IntentData>();
 
-            IntentData move3 = new IntentData(
+            IntentData move1 = new IntentData(
                     __instance.getClass(),
                     "BITE",
                     Intent.ATTACK
             );
-            move3.add_effect(new AttackEffect(__instance.damage.get(0)));
-            moves.add(move3);
+            move1.add_effect(new AttackEffect(__instance.damage.get(0)));
+            moves.add(move1);
 
-            IntentData move4 = new IntentData(
+            IntentData move2 = new IntentData(
                     __instance.getClass(),
-                    "STRENGTHEN",
+                    "GROW",
                     Intent.BUFF,
                     __instance.MOVES[0]
             );
+            int strAmt = (int)ReflectionHacks.getPrivate(__instance, __instance.getClass(), "strAmt");
             if (AbstractDungeon.ascensionLevel >= 17) {
-                move4.add_effect(new BuffStrengthEffect(4));
-            } else {
-                move4.add_effect(new BuffStrengthEffect(3));
+                strAmt += 1;
             }
-            moves.add(move4);
+            move2.add_effect(new BuffStrengthEffect(strAmt));
+            moves.add(move2);
 
             AbstractMonsterPatch.intent_moves.set(__instance, moves);
         }
