+++ src/main/java/chronometry/monsters/exordium/JawWormPatch.java
@@ -1,4 +1,4 @@
-package chronometry.monsters;
+package chronometry.monsters.exordium;
 
 import basemod.ReflectionHacks;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
