+++ src/main/java/chronometry/monsters/exordium/LouseNormalPatch.java
@@ -1,21 +1,22 @@
-package chronometry.monsters;
+package chronometry.monsters.exordium;
 
 import chronometry.patches.AbstractMonsterPatch;
 import chronometry.IntentData;
 import chronometry.effects.AttackEffect;
-import chronometry.effects.DebuffWeakEffect;
+import chronometry.effects.BuffStrengthEffect;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.monsters.AbstractMonster.Intent;
-import com.megacrit.cardcrawl.monsters.exordium.LouseDefensive;
+import com.megacrit.cardcrawl.monsters.exordium.LouseNormal;
 
 import java.util.ArrayList;
 
-public class LouseDefensivePatch {
-    @SpirePatch(clz=LouseDefensive.class,
+public class LouseNormalPatch {
+    @SpirePatch(clz=LouseNormal.class,
                 method=SpirePatch.CONSTRUCTOR,
                 paramtypez = {float.class, float.class})
     public static class InitMoves {
-        public static void Postfix(LouseDefensive __instance, float x, float y) {
+        public static void Postfix(LouseNormal __instance, float x, float y) {
             ArrayList<IntentData> moves = new ArrayList<IntentData>();
 
             IntentData move3 = new IntentData(
@@ -28,11 +29,15 @@ public class LouseDefensivePatch {
 
             IntentData move4 = new IntentData(
                     __instance.getClass(),
-                    "WEAKEN",
-                    Intent.DEBUFF,
+                    "STRENGTHEN",
+                    Intent.BUFF,
                     __instance.MOVES[0]
             );
-            move4.add_effect(new DebuffWeakEffect(2));
+            if (AbstractDungeon.ascensionLevel >= 17) {
+                move4.add_effect(new BuffStrengthEffect(4));
+            } else {
+                move4.add_effect(new BuffStrengthEffect(3));
+            }
             moves.add(move4);
 
             AbstractMonsterPatch.intent_moves.set(__instance, moves);
